+++ CHANGES.md
@@ -1,3 +1,4 @@
+- Integrated EJO Widget Template Loader
 - Allow capability for menu page to be filtered
 - Added Changelog